Refactor to remove need to run tests in a specific order
Now each test function can be run independently.
There is a tradeoff though. Test overlap. This was needed to remove the
need to run tests in specific order. Also Some tests are subset of
other. These are still kept so that, that part can be tested alone too.
For eg. issue creation is tested in almost all tests, but it is also
has a seperate test function, so that we can test issue creation
independently